refactor: merge duplicate code
Intersections can be sorted out by `unmap()`, so use that instead of having
two duplicate implementations for this, and encapsulate `merge()` as a
separate pass. This reduces the surface for possible bugs, and improves
test coverage at the same time, given that all `map()` tests now test also
`unmap()` code.

Closes: #40
Signed-off-by: Jarkko Sakkinen <jarkko@profian.com>