refactor: address range in `pub fn contains()`
In the public API,  replacing `(region)` with `(addr, length)` removes the
`region.start > region.end` corner case, and the need for check or assert
that. Therefore, use `(addr, length)` in the place of `(region)` in all
public functions.

In effect, this also simplifies call-sites, as the caller is not required
to create a region instance just to make a ledger call.

Closes: #52 ("refactor: get rid of counter-productive region.start < region.end checks")
Reported-by: Harald Hoyer <harald@profian.com>
Signed-off-by: Jarkko Sakkinen <jarkko@profian.com>